+++ apps/crawler/src/writers/writers.module.ts
@@ -1,0 +1,27 @@
+import { Block } from '@entities/Block';
+import { Collections } from '@entities/Collections';
+import { Tokens } from '@entities/Tokens';
+import { Event } from '@entities/Event';
+import { Module } from '@nestjs/common';
+import { ConfigModule } from '@nestjs/config';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Extrinsic } from '@entities/Extrinsic';
+import { Account } from '@entities/Account';
+import { CollectionWriterService } from './collection-writer.service';
+
+@Module({
+  imports: [
+    TypeOrmModule.forFeature([
+      Account,
+      Block,
+      Collections,
+      Event,
+      Extrinsic,
+      Tokens,
+    ]),
+    ConfigModule,
+  ],
+  providers: [CollectionWriterService],
+  exports: [CollectionWriterService],
+})
+export class WritersModule {}
